mail functionality completed
